Define `convertIdToMatrixId` method in tchaputils.TchqpPatterns
